+++ include/delaunay.hpp
@@ -296,7 +296,7 @@ namespace delaunay
 			double midy = (minY + maxY) * 0.5;
 
 			// Add our three super triangle vertices...
-			unsigned int p1Index = vertices_.size(), p2Index = vertices_.size() + 1, p3Index = vertices_.size() + 2;
+			unsigned int p1Index = static_cast<unsigned int>(vertices_.size()), p2Index = static_cast<unsigned int>(vertices_.size() + 1), p3Index = static_cast<unsigned int>(vertices_.size() + 2);
 
 			vector2 p1(midx - 20.0 * deltaMax, midy - deltaMax);
 			vector2 p2(midx, midy + 20.0 * deltaMax);
